Refaktoering og styling. Trekker UtledEllerVelgFagsak ut i egen fil, og legger til padding
